requirements.txt
altgraph==0.17.4
distlib==0.3.9
filelock==3.18.0
packaging==24.2
pefile==2023.2.7
platformdirs==4.3.7
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
python-dotenv==1.1.0
pywin32-ctypes==0.2.3
setuptools==78.1.0
virtualenv==20.29.3
